% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activateTask.R
\name{activateTask}
\alias{activateTask}
\title{Activate a task within a module in a pipeline}
\usage{
activateTask(pipeline, module, task)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{module}{the label of a module within \code{pipeline}}

\item{task}{a character string or vector. The label(s) of a task or tasks in \code{module} to activate.}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
activateTask(pipeline,module,task)
}

}

