% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTask.R
\name{addTask}
\alias{addTask}
\title{Function to create a new Task within a Module in an existing Pipeline}
\usage{
addTask(pipeline = NULL, module = NULL, label = NULL, method = NULL,
  datatype = "microarray", parameters = NULL, control = NULL,
  libraries = NULL)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{label}{a character string}

\item{type}{the type of module to add. One of 'M1', 'M2', or 'M3'}
}
\value{
function will execute silently
}
\description{
Author: Joe Perez-Rogers
}
\examples{

pipeline <- newPipeline(label="example pipeline", cv="cv", nfolds=10, p=0.80)
addModule(pipeline, type="M1", label="module 1")
addTask(pipeline, "module 1", label="task 1", method=function(x){return(x[1:5,])})
}

