% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_parquet.R
\name{convert_to_parquet}
\alias{convert_to_parquet}
\title{Save a CSV or TSV file in parquet format}
\usage{
convert_to_parquet(
  file_path,
  output_file = gsub(".csv", ".parquet", file_path, fixed = TRUE),
  ...
)
}
\arguments{
\item{file_path}{A string representing the path to the CSV file.}

\item{output_file}{path of output file}

\item{...}{Other arguments to pass into \code{read_csv_with_dates}}
}
\description{
Save a CSV or TSV file in parquet format
}
\examples{
\dontrun{
  df <- convert_to_parquet("data.csv")
}
}
