% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data.R
\name{crop_ends}
\alias{crop_ends}
\alias{crop_ends_diurnal}
\alias{crop_ends_nocturnal}
\title{Crop data to specified start and end times, to remove records at the beginning and end of each file.
This is useful for removing records collected before the device was attached or after it was removed. crop_diurnal defaults to cropping dataframe to the first instance of 01:00:00 and the last instance of 01:00:00. crop_nocturnal defaults to cropping dataframe to the first instance of 12:00:00 and the last instance of 12:00:00. crop_ends is a more general function that allows users to specify any start and end times. crop_diurnal}
\usage{
crop_ends(df, start_time = "00:00:00", end_time = "24:00:00")

crop_ends_diurnal(df, start_time = "01:00:00", end_time = "01:00:00")

crop_ends_nocturnal(df, start_time = "12:00:00", end_time = "12:00:00")
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{start_time}{The desired start time}

\item{end_time}{The desired end time}
}
\description{
Crop data to specified start and end times, to remove records at the beginning and end of each file.
This is useful for removing records collected before the device was attached or after it was removed. crop_diurnal defaults to cropping dataframe to the first instance of 01:00:00 and the last instance of 01:00:00. crop_nocturnal defaults to cropping dataframe to the first instance of 12:00:00 and the last instance of 12:00:00. crop_ends is a more general function that allows users to specify any start and end times. crop_diurnal
}
