% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_data.R
\name{crop_times}
\alias{crop_times}
\alias{crop_diurnal}
\alias{crop_nocturnal}
\title{This function Crops the start and end of each file to a specified time. Use crop_diurnal for diurnal data and crop_nocturnal for nocturnal data.}
\usage{
crop_times(df, start_time = "00:00:00", end_time = "24:00:00")

crop_diurnal(df, start_time = "06:00:00", end_time = "18:00:00")

crop_nocturnal(df, start_time = "18:00:00", end_time = "06:00:00")
}
\arguments{
\item{df}{A dataframe containing the data.}

\item{start_time}{The desired start time}

\item{end_time}{The desired end time}
}
\description{
This function Crops the start and end of each file to a specified time. Use crop_diurnal for diurnal data and crop_nocturnal for nocturnal data.
}
