% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deactivateTask.R
\name{deactivateTask}
\alias{deactivateTask}
\title{Deactivate a task within a module in a pipeline}
\usage{
deactivateTask(pipeline, module, task)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{module}{the label of a module within \code{pipeline}}

\item{task}{a character string or vector. The label(s) of a task or tasks in \code{module} to deactivate.}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
deactivateTask(pipeline,module,task)
}
}

