% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteModule.R
\name{deleteModule}
\alias{deleteModule}
\title{Delete a module from a pipeline}
\usage{
deleteModule(pipeline, module)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{module}{the label of a module within \code{pipeline}}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
deleteModule(pipeline,module)
}
}

