% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTask.R
\name{deleteTask}
\alias{deleteTask}
\title{Delete a task from a module within a pipeline}
\usage{
deleteTask(pipeline, module, task)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{module}{the label of a module within \code{pipeline}}

\item{task}{a character string or vector. The label(s) of a task or tasks in \code{module} to delete.}
}
\description{
A more detailed description will eventually go here...
}
\examples{

need examples
}

