% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jc2ad.R
\name{fetch_date_from_string}
\alias{fetch_date_from_string}
\title{fetch date from strng}
\usage{
fetch_date_from_string(input_str)
}
\arguments{
\item{input_str}{string}
}
\value{
Date
}
\description{
fetch date from strng
}
