% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCall.R
\name{getCall}
\alias{getCall}
\title{Get the function call used to add a task to a pipeline}
\usage{
getCall(pipeline, module, task)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{module}{the label of a module within \code{pipeline}}

\item{task}{a character string or vector. The label(s) of a task or tasks in \code{module} to delete.}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
getCall(pipeline,module,task)
}
}

