% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j2unicode.R
\name{jp2unicode}
\alias{jp2unicode}
\title{japanese charactor into escaped unicode}
\usage{
jp2unicode(str_vector)
}
\arguments{
\item{str_vector}{vector of strings}
}
\value{
vector of escaped unicode strings
}
\description{
japanese charactor into escaped unicode
}
