% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-class.R
\docType{class}
\name{Module}
\alias{Module}
\title{Class providing the Module object}
\format{\code{\link{R6Class}} object}
\usage{
Module
}
\value{
An object of class \code{Module}
}
\description{
Here is a more detailed sentence
about the Module class
}
\section{Fields}{

\describe{
\item{\code{label}}{The name of the Module}

\item{\code{class}}{The class of the Module (one of \code{M1}, \code{M2}, \code{M3}, \code{M4})}

\item{\code{tasks}}{A list of Task objects}

\item{\code{active}}{A Boolean vector of length \code{length(tasks)}}
}}
\section{Methods}{

\describe{
   \item{\code{addTask(label,method,parameters=NA,libraries=NA,control=NA,datatype)}}{Add a Task to the Module}
   \item{\code{deleteTask(label)}}{Remove a Task from the Module}
   \item{\code{run()}}{Execute a Task}
   \item{\code{activate(label)}}{Activate a Task}
   \item{\code{deactivate(label)}}{Deactivate a Task}
   \item{\code{summary()}}{Print out a summary of the Tasks in the Module}
}
}
\examples{


}
\keyword{biomarker}
\keyword{classification}

