% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{moving_window_calcs}
\alias{moving_window_calcs}
\alias{moving_window_calcs_2}
\title{moving_window_calcs}
\usage{
moving_window_calcs(df, window_size = 50)

moving_window_calcs_2(df, window_size = 50)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{window_size}{An integer specifying the size of the rolling window.}
}
\value{
A data frame with names following doAccloop.R
}
\description{
This function takes a data frame and computes sliding window calculations to set up for classification
}
