% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPipeline.R
\name{newPipeline}
\alias{newPipeline}
\title{Create an object of class Pipeline}
\usage{
newPipeline(label, cv = "lgocv", nfolds = 10, p = 0.8)
}
\arguments{
\item{label}{a character string}

\item{cv}{the cross-validation methodology to employ. One of 'cv', 'loocv', 'lgocv', or 'boot'}

\item{nfolds}{the number of folds to use in k-fold, leave-group out, or bootstrapped cross-validation}

\item{p}{the proportion of samples included in the training set in each fold}
}
\value{
an object of class \code{Pipeline}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
newPipeline(label="example pipeline",cv="lgocv",nfolds=10,p=0.80)
}
}

