% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderModules.R
\name{orderModules}
\alias{orderModules}
\title{Order the modules within a pipeline}
\usage{
orderModules(pipeline, order)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{order}{a character vector of module labels}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
orderModules(pipeline,order)
}
}

