% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline-class.R
\docType{class}
\name{Pipeline}
\alias{Pipeline}
\title{Class providing the Pipeline object}
\format{\code{\link{R6Class}} object}
\usage{
Pipeline$new(label, order = NULL, cv = "cv", nfolds = 10, p = NULL)
}
\value{
An object of class \code{Pipeline}
}
\description{
Here is a more detailed sentence
about the Pipeline class
}
\section{Fields}{

\describe{
\item{\code{label}}{The name of the Module}

\item{\code{order}}{An ordered character vector of \code{Module} names. \code{order} determined the order in which \code{Modules} will be run in the pipeline}

\item{\code{cv}}{The resampling method to be used for cross-validation. One of \code{"cv"}, \code{"loocv"}, \code{"lgocv"}, \code{"boot"}, or \code{"boot632"}}

\item{\code{nfolds}}{The number of folds for k-fold cross-validation or the number of resamplings for \code{"lgocv"}, \code{"boot"}, or \code{"boot632"}}

\item{\code{p}}{The percentage of samples in the internal training set for \code{"lgocv"}}
}}
\section{Methods}{

\describe{
   \item{\code{addModule(type,label)}}{Add a Module to the Pipeline}
   \item{\code{orderModules(x)}}{Change the order in which Modules are executed}
   \item{\code{run(x,y,data=NULL,ranks=NULL,verbose=FALSE,...)}}{Execute the pipeline}
}
}
\examples{

}
\keyword{biomarker}
\keyword{classification}

