% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbit-package.R
\docType{package}
\name{rabbit-package}
\alias{rabbit}
\alias{rabbit-package}
\title{rabbit: Rabbitly Process Data From Triaxial Accelerometers}
\description{
Rabbitly Process Data From Triaxial Accelerometers.
}
\author{
\strong{Maintainer}: Courntey Adams \email{courtney.adams@student.unsw.edu.au} (\href{https://orcid.org/0009-0007-3757-7507}{ORCID})

Authors:
\itemize{
  \item Jack Bilby [contributor]
  \item Daniel Falster (\href{https://orcid.org/0000-0002-9814-092X}{ORCID}) [contributor]
  \item Will Cornwell (\href{https://orcid.org/0000-0003-4080-4073}{ORCID}) [contributor]
}

}
\keyword{internal}
