% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_data.R
\name{standardise_data}
\alias{standardise_data}
\alias{standardize_data}
\title{Standardise accelerometer data}
\usage{
standardise_data(
  df,
  file_in = NULL,
  vars = c("Timestamp", "X", "Y", "Z"),
  time_function = lubridate::dmy_hms,
  timezone = "Australia/Adelaide",
  ...
)

standardize_data(
  df,
  file_in = NULL,
  vars = c("Timestamp", "X", "Y", "Z"),
  time_function = lubridate::dmy_hms,
  timezone = "Australia/Adelaide",
  ...
)
}
\arguments{
\item{df}{A data frame containing the data}

\item{file_in}{A string representing the path to the input file (CSV or parquet).}

\item{vars}{A character vector of length 4 specifying the names of the columns in df that contain the variables \code{time}, \code{X}, \code{Y}, \code{Z} from the  accelerometer. The default is c("Timestamp","X","Y","Z").}

\item{time_function}{A function to convert the time column to POSIXct. The default is lubridate::dmy_hms, which parses date-times with year, month, and day, hour, minute, and second components.}

\item{timezone}{Timezone of where accelerometer was used}

\item{...}{Other arguments to pass into read in function}
}
\value{
A data frame containing the standardised data, with the time column converted to POSIXct type.
}
\description{
This function accepts either a dataframe or filename (*.csv or *.parquet), checks that the time and x,y,z columns can be properly processed,
auto-detects common timestamp formats, and prints out summary statistics about the data.
}
\examples{
file_in = system.file("extdata", "raw_Pic2Jan_50000.parquet", package = "rabbit")
df <-
  standardise_data(file_in = file_in, vars = c("Timestamp", "accX", "accY", "accZ"))
}
