% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockPipeline.R
\docType{data}
\name{stockPipeline}
\alias{stockPipeline}
\title{Stock biomarker discovery pipeline for microarray data}
\format{An object of class \code{Pipeline}}
\usage{
data(stockPipeline)
}
\description{
More detail will go here
}
\examples{
data(stockPipeline)
}
\keyword{datasets}

