% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Pipeline}
\alias{summary.Pipeline}
\title{Report a summary of an object of class Pipeline}
\usage{
\method{summary}{Pipeline}(object, ...)
}
\arguments{
\item{object}{an object of class Pipeline}

\item{...}{additional parameters to be passed to summary. See Details below.}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
pipeline <- newPipeline(label="example pipeline",cv="lgocv",nfolds=10,p=0.80)
summary(pipeline)
}
}

