% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{summary_by_time}
\alias{summary_by_time}
\title{summary_by_time}
\usage{
summary_by_time(data, time_col, behavior_col, window_minutes)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{time_col}{time col}

\item{behavior_col}{behavior col}

\item{window_minutes}{how to chunk}
}
\value{
A data frame with a summary
}
\description{
summary_by_time
}
