% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task-class.R
\docType{class}
\name{Task}
\alias{Task}
\title{Class providing the Task object}
\format{\code{\link{R6Class}} object}
\usage{
Task
}
\value{
An object of class \code{Task} with members required to execute a discrete function
}
\description{
Here is a more detailed sentence
about the Task class
}
\section{Fields}{

\describe{
\item{\code{label}}{The name of the task}

\item{\code{method}}{A function to perform}

\item{\code{parameters}}{Additional input parameters used in \code{method}}

\item{\code{libraries}}{A character vector of libraries required to run the \code{method}}

\item{\code{control}}{A named list with values of \code{parameters} to pass to \code{method}}

\item{\code{datatype}}{A character vector of valid datatypes on which \code{method} can be used}
}}
\examples{


}
\keyword{biomarker}
\keyword{classification}

