% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update the configuration parameters of a Pipeline object}
\usage{
update(pipeline, what, value)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{what}{what to update? One of \code{label}, \code{cv}, \code{nfolds}, or \code{p}.}

\item{value}{value to update the variable named by \code{what}}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
update(pipeline=pipeline, what="cv", value="lgocv")
}

}

