% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-validateDatatype-method.R
\name{validateDatatype}
\alias{validateDatatype}
\title{Validate that input datatype conforms to required standards of the Task class}
\usage{
validateDatatype(datatype = NULL)
}
\arguments{
\item{label}{One or more of \code{microarray}, \code{count}, \code{rpkm}}
}
\value{
Function will execute silently if no errors are detected in \code{label}
}
\description{
More detailed explaination here
}
\examples{
validateDatatype("microarray")
validateDatatype(c("microarray","count"))
validateDatatype(c("microarray","count","rpkm"))
}

