% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-validateLabel-method.R
\name{validateLabel}
\alias{validateLabel}
\title{Validate that input label conforms to required standards of the Task class}
\usage{
validateLabel(label = NULL)
}
\arguments{
\item{label}{A character string}
}
\value{
Function will execute silently if no errors are detected in \code{label}
}
\description{
More detailed explaination here
}
\examples{
validateLabel("label")

}

