% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-validateLibraries-method.R
\name{validateLibraries}
\alias{validateLibraries}
\title{Validate that input libraries conform to required standards of the Task class}
\usage{
validateLibraries(libraries = NULL)
}
\arguments{
\item{libraries}{A character vector of libraries}
}
\value{
Function will execute silently if no errors are detected in \code{libraries}
}
\description{
More detailed explaination here
}
\examples{
validateLibraries("limma")
validateDatatype(c("limma","caret"))
}

