% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateParameters.R
\name{validateParameters}
\alias{validateParameters}
\title{Validate that a data.frame of parameters conforms to required standards of the Task class}
\usage{
validateParameters(parameters)
}
\arguments{
\item{parameters}{A data.frame with columns \code{parameter}, \code{class}, and \code{label}}
}
\value{
TRUE if no errors are detected in \code{parameters}
}
\description{
More detailed explaination here
}
\examples{
parameters <- data.frame(
 parameter = c("x","y"),
 class = c("logical","numeric"),
 label = c("myX","myY")
 )
validateParameters(parameters)
}

