% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv.R
\name{wv.model}
\alias{wv.model}
\title{Train a weighted voting classifier}
\usage{
wv.model(data, classlabel, correction = TRUE)
}
\arguments{
\item{data}{matrix with samples in columns and genes in rows}

\item{classlabel}{two-level factor of binary outcomes (0 or 1)}

\item{correction}{should the Broad Institute's correction method be used?}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
library(multtest)
data(golub)
mod <- wv.model(golub,golub.cl,correction=T)
}
}

