% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv.predict.R
\name{wv.predict}
\alias{wv.predict}
\title{Train a weighted voting classifier}
\usage{
wv.predict(model, data, type = "raw")
}
\arguments{
\item{model}{a list as generated by wv.model()}

\item{data}{matrix with samples in columns and genes in rows}

\item{type}{one of "raw" or "prob"}
}
\description{
A more detailed description will eventually go here...
}
\examples{
\dontrun{
library(multtest)
data(golub)
mod <- wv.model(golub,golub.cl,correction=T)
wv.predict(mod,golub,type="raw")
}
}

