% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addModule.R
\name{addModule}
\alias{addModule}
\title{Create a new Module within an existing Pipeline}
\usage{
addModule(pipeline, type, label)
}
\arguments{
\item{pipeline}{an object of class \code{Pipeline}}

\item{type}{the type of module to add. One of 'M1', 'M2', or 'M3'}

\item{label}{a character string}
}
\description{
A more detailed description will eventually go here...
}
\examples{

pipeline <- newPipeline(label="example pipeline",cv="cv",nfolds=10,p=0.80)
addModule(pipeline,type="M1",label="module 1")
}

