% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_behaviors}
\alias{classify_behaviors}
\title{classify_behaviors}
\usage{
classify_behaviors(
  dat,
  MSOM_path = "tests/testthat/MSOM_8by7.rda",
  quiet = FALSE
)
}
\arguments{
\item{dat}{A data frame containing the data.}

\item{MSOM_path}{classifier}

\item{quiet}{chill out}
}
\value{
A data frame with a column of behaviors
}
\description{
This function takes a data frame and a classifier and predicts across the data frame assuming everything lines up
}
