% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_parquet.R
\name{convert_to_parquet}
\alias{convert_to_parquet}
\title{Save a CSV or TSV file in parquet format}
\usage{
convert_to_parquet(
  file,
  output_file = gsub(".csv", ".parquet", file, fixed = TRUE),
  delim = "\\t"
)
}
\arguments{
\item{file}{path of input file}

\item{output_file}{path of output file}

\item{delim}{delimiter (default = "\t")}
}
\description{
Save a CSV or TSV file in parquet format
}
