% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_movement_dynamics.R
\name{extract_movement_dynamics}
\alias{extract_movement_dynamics}
\title{This function takes standardised accelerometer data and caluclates
a set of metrics of movement dynamics across a sliding window of specified number of datapoints. Caluclate metrics include including means, variances, covariances, of xyz, as well as overall dynamic body acceleration (ODBA) and vectorial dynamic body acceleration (VDBA). The function is designed to be fast, using the RcppRoll package to calculate rolling means and sums in C++.}
\usage{
extract_movement_dynamics(df, window_size = 50)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{window_size}{An integer specifying the size of the rolling window.}
}
\value{
A data frame with movement dynamics metrics calculated across the sliding window.
}
\description{
This function takes standardised accelerometer data and caluclates
a set of metrics of movement dynamics across a sliding window of specified number of datapoints. Caluclate metrics include including means, variances, covariances, of xyz, as well as overall dynamic body acceleration (ODBA) and vectorial dynamic body acceleration (VDBA). The function is designed to be fast, using the RcppRoll package to calculate rolling means and sums in C++.
}
\examples{
#' @examples
file_in = system.file("extdata", "raw_Pic2Jan_50000.parquet", package = "rabbit")
df <-
  standardize_data(file_in = file_in, vars = c("Timestamp", "accX", "accY", "accZ")) |>
  extract_movement_dynamics()
}
