% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPipeline.R
\name{newPipeline}
\alias{newPipeline}
\title{Function to create a new instance of a Pipeline}
\usage{
newPipeline(label = NULL, cv = "cv", nfolds = 10, p = 0.8)
}
\arguments{
\item{label}{a character string}

\item{cv}{the cross-validation methodology to employ. One of 'cv', 'loocv', 'lgocv', or 'boot'}

\item{nfolds}{the number of folds to use in k-fold, leave-group out, or bootstrapped cross-validation}

\item{p}{the proportion of samples included in the training set in each fold}
}
\value{
an object of class \code{Pipeline}
}
\description{
Author: Joe Perez-Rogers
}
\examples{

newPipeline(label="example pipeline",cv="cv",nfolds=10,p=0.80)
}

