% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\alias{standardise_data}
\title{Standardize accelerometer data}
\usage{
standardize_data(
  df,
  file_in = NULL,
  file_out = NULL,
  vars = c("Timestamp", "X", "Y", "Z"),
  timezone = "Australia/Adelaide",
  ...
)

standardise_data(
  df,
  file_in = NULL,
  file_out = NULL,
  vars = c("Timestamp", "X", "Y", "Z"),
  timezone = "Australia/Adelaide",
  ...
)
}
\arguments{
\item{df}{A data frame containing the data}

\item{file_in}{A string representing the path to the input file (CSV or parquet).}

\item{file_out}{path where the output file should be saved. If NULL, the data will not be saved to a file. The output format will be determined by the file extension (CSV or parquet).}

\item{vars}{A character vector of length 4 specifying the names of the columns in df that contain the variables \code{time}, \code{X}, \code{Y}, \code{Z} from the  accelerometer. The default is c("Timestamp","X","Y","Z").}

\item{timezone}{Timezone of where accelerometer was used}

\item{...}{Other arguments to pass into read in function}
}
\value{
A data frame containing the standardised data, with the time column converted to POSIXct type.
}
\description{
This function accpets either a dataframe or filename (*.csv or *.parquet), checks that the time and x,y,z columncan be properly processed,
and prints out summary statistics about the data.
}
\examples{
file_in = system.file("extdata", "raw_Pic2Jan_50000.parquet", package = "rabbit")
df <-
  standardize_data(file_in = file_in, vars = c("Timestamp", "accX", "accY", "accZ"))
}
