% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stockPipeline.R
\docType{data}
\name{stockPipeline}
\alias{stockPipeline}
\title{Stock implementation of a biomarker discovery pipeline}
\format{a \code{Pipeline} object containing 4 modules and 15 total tasks.}
\usage{
stockPipeline
}
\description{
A more detailed description will eventually go here...
}
\examples{
summary(stockPipeline, level="structure")
}
\keyword{datasets}

