% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.Pipeline}
\alias{summary.Pipeline}
\title{Report a summary of an object of class Pipeline}
\usage{
\method{summary}{Pipeline}(x, level = "overview", pretty = TRUE)
}
\arguments{
\item{x}{an object of class Pipeline}

\item{level}{one of 'overview', 'structure', or 'active' (see details below)}
}
\description{
A more detailed description will eventually go here...
}
\examples{

pipeline <- newPipeline(label="example pipeline",cv="cv",nfolds=10,p=0.80)
summary(pipeline)
}

